move slatedb/src/dst back to slatedb-dst
