--- slatedb-dst/src/error.rs
@@ -1,17 +1,0 @@
-use slatedb::Error;
-use thiserror::Error as ThisError;
-
-#[derive(Debug, ThisError)]
-pub(crate) enum DstError {
-    #[error("failed to perform an operation on the SQLite state database")]
-    SQLiteStateError(#[from] rusqlite::Error),
-}
-
-impl From<DstError> for Error {
-    fn from(err: DstError) -> Self {
-        let msg = err.to_string();
-        match err {
-            DstError::SQLiteStateError(e) => Error::internal(msg).with_source(Box::new(e)),
-        }
-    }
-}
