+++ slatedb-dst/src/failing_object_store.rs
@@ -14,9 +14,8 @@ use object_store::{
 };
 use parking_lot::RwLock;
 use rand::RngCore;
+use slatedb::DbRand;
 use slatedb_common::clock::SystemClock;
-
-use crate::rand::DbRand;
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub enum StreamDirection {
@@ -259,32 +258,17 @@ impl FailingObjectStore {
 
     async fn put_like(
         &self,
-        operation: Operation,
-        from: &Path,
-        to: Option<&Path>,
-        request_size: u64,
-        inner: impl std::future::Future<Output = object_store::Result<PutResult>>,
+        location: &Path,
+        payload: PutPayload,
+        opts: PutOptions,
     ) -> object_store::Result<PutResult> {
-        let paths = collect_paths(from, to);
-        self.apply_request_faults(operation, &paths, request_size)
+        let size = u64::try_from(payload.content_length()).unwrap_or(u64::MAX);
+        self.apply_request_faults(Operation::PutOpts, &[location.clone()], size)
             .await?;
-        let result = inner.await?;
-        self.apply_response_faults(operation, &paths, 0).await?;
+        let result = self.inner.put_opts(location, payload, opts).await?;
+        self.apply_response_faults(Operation::PutOpts, &[location.clone()], 0)
+            .await?;
         Ok(result)
-    }
-
-    async fn unit_result(
-        &self,
-        operation: Operation,
-        from: &Path,
-        to: Option<&Path>,
-        inner: impl std::future::Future<Output = object_store::Result<()>>,
-    ) -> object_store::Result<()> {
-        let paths = collect_paths(from, to);
-        self.apply_request_faults(operation, &paths, 0).await?;
-        inner.await?;
-        self.apply_response_faults(operation, &paths, 0).await?;
-        Ok(())
     }
 }
 
@@ -302,30 +286,60 @@ impl fmt::Display for FailingObjectStore {
 
 #[async_trait]
 impl ObjectStore for FailingObjectStore {
+    async fn put_opts(
+        &self,
+        location: &Path,
+        payload: PutPayload,
+        opts: PutOptions,
+    ) -> object_store::Result<PutResult> {
+        self.put_like(location, payload, opts).await
+    }
+
+    async fn put_multipart(
+        &self,
+        location: &Path,
+    ) -> object_store::Result<Box<dyn MultipartUpload>> {
+        self.apply_request_faults(Operation::PutOpts, &[location.clone()], 0)
+            .await?;
+        let result = self.inner.put_multipart(location).await?;
+        self.apply_response_faults(Operation::PutOpts, &[location.clone()], 0)
+            .await?;
+        Ok(result)
+    }
+
+    async fn put_multipart_opts(
+        &self,
+        location: &Path,
+        opts: PutMultipartOptions,
+    ) -> object_store::Result<Box<dyn MultipartUpload>> {
+        self.apply_request_faults(Operation::PutOpts, &[location.clone()], 0)
+            .await?;
+        let result = self.inner.put_multipart_opts(location, opts).await?;
+        self.apply_response_faults(Operation::PutOpts, &[location.clone()], 0)
+            .await?;
+        Ok(result)
+    }
+
     async fn get_opts(
         &self,
         location: &Path,
         options: GetOptions,
     ) -> object_store::Result<GetResult> {
-        let paths = vec![location.clone()];
-        self.apply_request_faults(Operation::GetOpts, &paths, 0)
+        self.apply_request_faults(Operation::GetOpts, &[location.clone()], 0)
             .await?;
         let result = self.inner.get_opts(location, options).await?;
-        self.apply_response_faults(
-            Operation::GetOpts,
-            &paths,
-            result.range.end.saturating_sub(result.range.start),
-        )
-        .await?;
+        let size = result.meta.size;
+        self.apply_response_faults(Operation::GetOpts, &[location.clone()], size)
+            .await?;
         Ok(result)
     }
 
     async fn get_range(&self, location: &Path, range: Range<u64>) -> object_store::Result<Bytes> {
-        let paths = vec![location.clone()];
-        self.apply_request_faults(Operation::GetRange, &paths, 0)
+        self.apply_request_faults(Operation::GetRange, &[location.clone()], 0)
             .await?;
         let result = self.inner.get_range(location, range).await?;
-        self.apply_response_faults(Operation::GetRange, &paths, result.len() as u64)
+        let size = u64::try_from(result.len()).unwrap_or(u64::MAX);
+        self.apply_response_faults(Operation::GetRange, &[location.clone()], size)
             .await?;
         Ok(result)
     }
@@ -335,127 +349,63 @@ impl ObjectStore for FailingObjectStore {
         location: &Path,
         ranges: &[Range<u64>],
     ) -> object_store::Result<Vec<Bytes>> {
-        let paths = vec![location.clone()];
-        self.apply_request_faults(Operation::GetRanges, &paths, 0)
+        self.apply_request_faults(Operation::GetRanges, &[location.clone()], 0)
             .await?;
         let result = self.inner.get_ranges(location, ranges).await?;
-        let total_size = result.iter().map(|bytes| bytes.len() as u64).sum();
-        self.apply_response_faults(Operation::GetRanges, &paths, total_size)
+        let size = result.iter().fold(0_u64, |total, bytes| {
+            total.saturating_add(u64::try_from(bytes.len()).unwrap_or(u64::MAX))
+        });
+        self.apply_response_faults(Operation::GetRanges, &[location.clone()], size)
             .await?;
         Ok(result)
     }
 
     async fn head(&self, location: &Path) -> object_store::Result<ObjectMeta> {
-        let paths = vec![location.clone()];
-        self.apply_request_faults(Operation::Head, &paths, 0)
+        self.apply_request_faults(Operation::Head, &[location.clone()], 0)
             .await?;
         let result = self.inner.head(location).await?;
-        self.apply_response_faults(Operation::Head, &paths, 0)
+        self.apply_response_faults(Operation::Head, &[location.clone()], result.size)
             .await?;
         Ok(result)
     }
 
-    async fn put_opts(
-        &self,
-        location: &Path,
-        payload: PutPayload,
-        opts: PutOptions,
-    ) -> object_store::Result<PutResult> {
-        self.put_like(
-            Operation::PutOpts,
-            location,
-            None,
-            payload.content_length() as u64,
-            self.inner.put_opts(location, payload, opts),
-        )
-        .await
-    }
-
-    async fn put_multipart(
-        &self,
-        location: &Path,
-    ) -> object_store::Result<Box<dyn MultipartUpload>> {
-        self.inner.put_multipart(location).await
-    }
-
-    async fn put_multipart_opts(
-        &self,
-        location: &Path,
-        opts: PutMultipartOptions,
-    ) -> object_store::Result<Box<dyn MultipartUpload>> {
-        self.inner.put_multipart_opts(location, opts).await
-    }
-
     async fn delete(&self, location: &Path) -> object_store::Result<()> {
-        self.unit_result(
-            Operation::Delete,
-            location,
-            None,
-            self.inner.delete(location),
-        )
-        .await
+        self.apply_request_faults(Operation::Delete, &[location.clone()], 0)
+            .await?;
+        let result = self.inner.delete(location).await;
+        self.apply_response_faults(Operation::Delete, &[location.clone()], 0)
+            .await?;
+        result
     }
 
     fn list(&self, prefix: Option<&Path>) -> BoxStream<'static, object_store::Result<ObjectMeta>> {
-        let inner = Arc::clone(&self.inner);
         let this = self.clone();
+        let this_for_response = self.clone();
+        let inner = Arc::clone(&self.inner);
         let prefix_owned = prefix.cloned();
+        let request_path = prefix_owned.clone().unwrap_or_default();
+        let response_path = request_path.clone();
         stream::once(async move {
-            let paths = prefix_owned.iter().cloned().collect::<Vec<_>>();
-            this.apply_request_faults(Operation::List, &paths, 0)
-                .await?;
-
-            let downstream = this.controller.sample_matching_toxics(
-                StreamDirection::Downstream,
-                Operation::List,
-                &paths,
-            );
-            if downstream
-                .iter()
-                .any(|toxic| matches!(toxic, ToxicKind::ResetPeer))
-            {
-                let error_stream: BoxStream<'static, object_store::Result<ObjectMeta>> =
-                    stream::once(async { Err(reset_peer_error()) }).boxed();
-                return Ok::<_, object_store::Error>(error_stream);
-            }
-
-            let pre_return = downstream
-                .iter()
-                .filter(|toxic| !matches!(toxic, ToxicKind::Bandwidth { .. }))
-                .cloned()
-                .collect::<Vec<_>>();
-            this.apply_toxics(&pre_return, 0, true).await?;
-
-            let bandwidth = downstream
-                .into_iter()
-                .filter_map(|toxic| match toxic {
-                    ToxicKind::Bandwidth { bytes_per_sec } => Some(bytes_per_sec),
-                    _ => None,
-                })
-                .collect::<Vec<_>>();
-
-            let clock = Arc::clone(&this.clock);
-            let stream = inner
-                .list(prefix_owned.as_ref())
-                .then(move |result| {
-                    let clock = Arc::clone(&clock);
-                    let bandwidth = bandwidth.clone();
-                    async move {
-                        if let Ok(meta) = &result {
-                            let size = estimate_meta_size(meta);
-                            for bytes_per_sec in &bandwidth {
-                                if let Some(delay) = bandwidth_delay(size, *bytes_per_sec) {
-                                    clock.advance(delay).await;
-                                }
-                            }
-                        }
-                        result
-                    }
-                })
-                .boxed();
-            Ok(stream)
+            this.apply_request_faults(Operation::List, std::slice::from_ref(&request_path), 0)
+                .await
         })
+        .map_ok(move |_| inner.list(prefix_owned.as_ref()))
         .try_flatten()
+        .then(move |result| {
+            let this = this_for_response.clone();
+            let response_path = response_path.clone();
+            async move {
+                if let Ok(meta) = &result {
+                    this.apply_response_faults(
+                        Operation::List,
+                        std::slice::from_ref(&response_path),
+                        meta.size,
+                    )
+                    .await?;
+                }
+                result
+            }
+        })
         .boxed()
     }
 
@@ -464,147 +414,116 @@ impl ObjectStore for FailingObjectStore {
         prefix: Option<&Path>,
         offset: &Path,
     ) -> BoxStream<'static, object_store::Result<ObjectMeta>> {
-        let inner = Arc::clone(&self.inner);
         let this = self.clone();
+        let this_for_response = self.clone();
+        let inner = Arc::clone(&self.inner);
         let prefix_owned = prefix.cloned();
         let offset_owned = offset.clone();
+        let request_paths = vec![
+            prefix_owned.clone().unwrap_or_default(),
+            offset_owned.clone(),
+        ];
+        let response_paths = request_paths.clone();
         stream::once(async move {
-            let paths = collect_optional_paths(prefix_owned.as_ref(), Some(&offset_owned));
-            this.apply_request_faults(Operation::ListWithOffset, &paths, 0)
-                .await?;
-
-            let downstream = this.controller.sample_matching_toxics(
-                StreamDirection::Downstream,
-                Operation::ListWithOffset,
-                &paths,
-            );
-            if downstream
-                .iter()
-                .any(|toxic| matches!(toxic, ToxicKind::ResetPeer))
-            {
-                let error_stream: BoxStream<'static, object_store::Result<ObjectMeta>> =
-                    stream::once(async { Err(reset_peer_error()) }).boxed();
-                return Ok::<_, object_store::Error>(error_stream);
-            }
-
-            let pre_return = downstream
-                .iter()
-                .filter(|toxic| !matches!(toxic, ToxicKind::Bandwidth { .. }))
-                .cloned()
-                .collect::<Vec<_>>();
-            this.apply_toxics(&pre_return, 0, true).await?;
-
-            let bandwidth = downstream
-                .into_iter()
-                .filter_map(|toxic| match toxic {
-                    ToxicKind::Bandwidth { bytes_per_sec } => Some(bytes_per_sec),
-                    _ => None,
-                })
-                .collect::<Vec<_>>();
-
-            let clock = Arc::clone(&this.clock);
-            let stream = inner
-                .list_with_offset(prefix_owned.as_ref(), &offset_owned)
-                .then(move |result| {
-                    let clock = Arc::clone(&clock);
-                    let bandwidth = bandwidth.clone();
-                    async move {
-                        if let Ok(meta) = &result {
-                            let size = estimate_meta_size(meta);
-                            for bytes_per_sec in &bandwidth {
-                                if let Some(delay) = bandwidth_delay(size, *bytes_per_sec) {
-                                    clock.advance(delay).await;
-                                }
-                            }
-                        }
-                        result
-                    }
-                })
-                .boxed();
-            Ok(stream)
+            this.apply_request_faults(Operation::ListWithOffset, &request_paths, 0)
+                .await
         })
+        .map_ok(move |_| inner.list_with_offset(prefix_owned.as_ref(), &offset_owned))
         .try_flatten()
+        .then(move |result| {
+            let this = this_for_response.clone();
+            let response_paths = response_paths.clone();
+            async move {
+                if let Ok(meta) = &result {
+                    this.apply_response_faults(
+                        Operation::ListWithOffset,
+                        &response_paths,
+                        meta.size,
+                    )
+                    .await?;
+                }
+                result
+            }
+        })
         .boxed()
     }
 
     async fn list_with_delimiter(&self, prefix: Option<&Path>) -> object_store::Result<ListResult> {
-        self.inner.list_with_delimiter(prefix).await
+        let path = prefix.cloned().unwrap_or_default();
+        self.apply_request_faults(Operation::List, std::slice::from_ref(&path), 0)
+            .await?;
+        let result = self.inner.list_with_delimiter(prefix).await?;
+        let size = result
+            .objects
+            .iter()
+            .fold(0_u64, |total, meta| total.saturating_add(meta.size));
+        self.apply_response_faults(Operation::List, &[path], size)
+            .await?;
+        Ok(result)
     }
 
     async fn copy(&self, from: &Path, to: &Path) -> object_store::Result<()> {
-        self.unit_result(Operation::Copy, from, Some(to), self.inner.copy(from, to))
+        self.apply_request_faults(Operation::Copy, &[from.clone(), to.clone()], 0)
+            .await?;
+        self.inner.copy(from, to).await?;
+        self.apply_response_faults(Operation::Copy, &[from.clone(), to.clone()], 0)
             .await
     }
 
     async fn rename(&self, from: &Path, to: &Path) -> object_store::Result<()> {
-        self.unit_result(
-            Operation::Rename,
-            from,
-            Some(to),
-            self.inner.rename(from, to),
-        )
-        .await
+        self.apply_request_faults(Operation::Rename, &[from.clone(), to.clone()], 0)
+            .await?;
+        self.inner.rename(from, to).await?;
+        self.apply_response_faults(Operation::Rename, &[from.clone(), to.clone()], 0)
+            .await
     }
 
     async fn copy_if_not_exists(&self, from: &Path, to: &Path) -> object_store::Result<()> {
-        self.unit_result(
-            Operation::Copy,
-            from,
-            Some(to),
-            self.inner.copy_if_not_exists(from, to),
-        )
-        .await
+        self.apply_request_faults(Operation::Copy, &[from.clone(), to.clone()], 0)
+            .await?;
+        self.inner.copy_if_not_exists(from, to).await?;
+        self.apply_response_faults(Operation::Copy, &[from.clone(), to.clone()], 0)
+            .await
     }
 
     async fn rename_if_not_exists(&self, from: &Path, to: &Path) -> object_store::Result<()> {
-        self.unit_result(
-            Operation::Rename,
-            from,
-            Some(to),
-            self.inner.rename_if_not_exists(from, to),
-        )
-        .await
+        self.apply_request_faults(Operation::Rename, &[from.clone(), to.clone()], 0)
+            .await?;
+        self.inner.rename_if_not_exists(from, to).await?;
+        self.apply_response_faults(Operation::Rename, &[from.clone(), to.clone()], 0)
+            .await
     }
 }
 
 #[derive(Debug)]
 pub struct HttpStatusError {
-    pub status_code: u16,
+    status_code: u16,
+}
+
+impl HttpStatusError {
+    pub fn status_code(&self) -> u16 {
+        self.status_code
+    }
 }
 
 impl fmt::Display for HttpStatusError {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "injected HTTP status {}", self.status_code)
+        write!(
+            f,
+            "HTTP error before request dispatch: {}",
+            self.status_code
+        )
     }
 }
 
 impl std::error::Error for HttpStatusError {}
 
-fn collect_paths(primary: &Path, secondary: Option<&Path>) -> Vec<Path> {
-    let mut paths = vec![primary.clone()];
-    if let Some(secondary) = secondary {
-        paths.push(secondary.clone());
-    }
-    paths
-}
-
-fn collect_optional_paths(primary: Option<&Path>, secondary: Option<&Path>) -> Vec<Path> {
-    let mut paths = Vec::new();
-    if let Some(primary) = primary {
-        paths.push(primary.clone());
-    }
-    if let Some(secondary) = secondary {
-        paths.push(secondary.clone());
-    }
-    paths
-}
-
 fn path_matches(prefix: &Option<String>, paths: &[Path]) -> bool {
     match prefix {
+        None => true,
         Some(prefix) => paths
             .iter()
-            .any(|path| path.to_string().starts_with(prefix)),
-        None => true,
+            .any(|path| path.as_ref().starts_with(prefix.as_str())),
     }
 }
 
@@ -613,28 +532,16 @@ fn bandwidth_delay(size_bytes: u64, bytes_per_sec: u64) -> Option<Duration> {
         return None;
     }
 
-    let millis = ((size_bytes as u128) * 1000).div_ceil(bytes_per_sec as u128);
-    Some(Duration::from_millis(
-        u64::try_from(millis).unwrap_or(u64::MAX),
-    ))
-}
-
-fn estimate_meta_size(meta: &ObjectMeta) -> u64 {
-    let mut size = meta.location.to_string().len() as u64;
-    size += meta.e_tag.as_ref().map_or(0, |tag| tag.len() as u64);
-    size += meta
-        .version
-        .as_ref()
-        .map_or(0, |version| version.len() as u64);
-    size + 32
+    let millis = size_bytes.saturating_mul(1_000).div_ceil(bytes_per_sec);
+    Some(Duration::from_millis(millis.max(1)))
 }
 
 fn reset_peer_error() -> object_store::Error {
     object_store::Error::Generic {
-        store: "dst_reset_peer",
+        store: "dst_toxic_reset_peer",
         source: Box::new(std::io::Error::new(
             std::io::ErrorKind::ConnectionReset,
-            "dst toxic reset_peer injected a transport reset",
+            "dst toxic reset peer",
         )),
     }
 }
